Splitting Default service into multiple services and adding documentation
