+++ controllers/Default.js
@@ -1,10 +1,23 @@
+/**
+ * @file Default.js
+ * @description This file contains the controller functions for the API endpoints.
+ * Each function corresponds to an endpoint and interacts with the DefaultService to perform the required operations.
+ * The responses are written using the utils.writeJson method.
+ */
+
 'use strict';
 
 var utils = require('../utils/writer.js');
-var Default = require('../service/DefaultService');
+var GetService = require('../service/GetService');
+var PostService = require('../service/PostService');
+var PutService = require('../service/PutService');
+var DeleteService = require('../service/DeleteService');
+var AlgorithmicService = require('../service/AlgorithmicService');
 
+
+// Add accommodation to a trip
 module.exports.addAccommodation = function addAccommodation (req, res, next, body, userId, tripId) {
-  Default.addAccommodation(body, userId, tripId)
+  PostService.addAccommodation(body, userId, tripId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -13,8 +26,9 @@ module.exports.addAccommodation = function addAccommodation (req, res, next, bod
     });
 };
 
+// Add activity to a day
 module.exports.addActivity = function addActivity (req, res, next, body, userId, tripId, dayId) {
-  Default.addActivity(body, userId, tripId, dayId)
+  PostService.addActivity(body, userId, tripId, dayId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -23,8 +37,9 @@ module.exports.addActivity = function addActivity (req, res, next, body, userId,
     });
 };
 
+// Add day to a trip
 module.exports.addDay = function addDay (req, res, next, userId, tripId) {
-  Default.addDay(userId, tripId)
+  PostService.addDay(userId, tripId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -33,8 +48,9 @@ module.exports.addDay = function addDay (req, res, next, userId, tripId) {
     });
 };
 
+// Add transportation to a trip
 module.exports.addTransportation = function addTransportation (req, res, next, body, userId, tripId) {
-  Default.addTransportation(body, userId, tripId)
+  PostService.addTransportation(body, userId, tripId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -43,8 +59,9 @@ module.exports.addTransportation = function addTransportation (req, res, next, b
     });
 };
 
+// Create a new trip
 module.exports.createTrip = function createTrip (req, res, next, body, userId) {
-  Default.createTrip(body, userId)
+  PostService.createTrip(body, userId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -53,8 +70,9 @@ module.exports.createTrip = function createTrip (req, res, next, body, userId) {
     });
 };
 
+// Create a new user
 module.exports.createUser = function createUser (req, res, next, body) {
-  Default.createUser(body)
+  PostService.createUser(body)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -63,8 +81,9 @@ module.exports.createUser = function createUser (req, res, next, body) {
     });
 };
 
+// Delete an activity
 module.exports.deleteActivity = function deleteActivity (req, res, next, userId, tripId, dayId, activityId) {
-  Default.deleteActivity(userId, tripId, dayId, activityId)
+  DeleteService.deleteActivity(userId, tripId, dayId, activityId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -73,8 +92,9 @@ module.exports.deleteActivity = function deleteActivity (req, res, next, userId,
     });
 };
 
+// Generate random activities for a certain day
 module.exports.generateRandomActivities = function generateRandomActivities (req, res, next, userId, tripId, dayId) {
-  Default.generateRandomActivities(userId, tripId, dayId)
+  AlgorithmicService.generateRandomActivities(userId, tripId, dayId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -83,8 +103,9 @@ module.exports.generateRandomActivities = function generateRandomActivities (req
     });
 };
 
+// Get an activity
 module.exports.getActivity = function getActivity (req, res, next, userId, tripId, dayId, activityId) {
-  Default.getActivity(userId, tripId, dayId, activityId)
+  GetService.getActivity(userId, tripId, dayId, activityId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -93,8 +114,9 @@ module.exports.getActivity = function getActivity (req, res, next, userId, tripI
     });
 };
 
+// Get a day
 module.exports.getDay = function getDay (req, res, next, userId, tripId, dayId) {
-  Default.getDay(userId, tripId, dayId)
+  GetService.getDay(userId, tripId, dayId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -103,8 +125,9 @@ module.exports.getDay = function getDay (req, res, next, userId, tripId, dayId) 
     });
 };
 
+// Get a trip
 module.exports.getTrip = function getTrip (req, res, next, userId, tripId) {
-  Default.getTrip(userId, tripId)
+  GetService.getTrip(userId, tripId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -113,8 +136,9 @@ module.exports.getTrip = function getTrip (req, res, next, userId, tripId) {
     });
 };
 
+// Get all trips of a user
 module.exports.getUsersTrips = function getUsersTrips (req, res, next, userId) {
-  Default.getUsersTrips(userId)
+  GetService.getUsersTrips(userId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -123,8 +147,9 @@ module.exports.getUsersTrips = function getUsersTrips (req, res, next, userId) {
     });
 };
 
+// Update an activity
 module.exports.updateActivity = function updateActivity (req, res, next, body, userId, tripId, dayId, activityId) {
-  Default.updateActivity(body, userId, tripId, dayId, activityId)
+  PutService.updateActivity(body, userId, tripId, dayId, activityId)
     .then(function (response) {
       utils.writeJson(res, response);
     })
